tessl: audit, hardening, and measured kernel optimizations
Everything that was uncommitted under crates/tessl, the tessl CI workflow,
and the three metal-native call sites that the bump_reset signature change
requires (rt.bump_reset()? in engine.rs and train.rs).

The 2026-09-04 audit (three reports: NN kernels, GEMM, runtime; 45 items)
and its fixes, each with a test that fails against the pre-fix code:
q-only RoPE re-normalizing rows other threads owned; Interleaved4 Q4 banks
validated at their tile-padded extent; Q4 group sizes and column counts the
kernels cannot address refused; double softcap in multi-pass argmax;
cross-scope producer/consumer edges in hazard-barrier mode, carried by the
runtime and inserted lazily (Binder::new takes a BarrierPolicy); split-K
partitions planned and capped at 32; launch geometry that assumed a 32-wide
simdgroup now refuses anything else. Robustness: a failed or panicking
encode closure ends its encoder and command buffer and discards the batch;
bump_reset returns Result; bump views read and write only their own window;
the Metal 4 state machine is crate-private; build.rs prints the metal4
diagnostic on fallback; the dispatch counter counts every attempt.

Measured (paired, order-balanced, in-process A/B, twenty dispatches per
command buffer, M5 Pro): the fused RMSNorm+QKV+RoPE kernels are one
simdgroup per head row, 3.4-11.5x at decode shapes and 1.3-1.7x at long
prefill (bench/results/rope_simdgroup_m5pro.txt); the NN coop GEMM tile is
chosen by grid fill, 1.35-1.74x at M <= 64 and 1.14-1.59x on starved grids
(bench/results/bf16_smallm_coop_m5pro.txt); host access with no GPU work
pending no longer commits residency; crate-internal temporaries skip the
host memset.

Also carried from the earlier uncommitted work on this crate, as recorded
in CHANGELOG.md [Unreleased]: the nn module promoted out of gemma-metal,
int8 and strided-batched GEMM, f16 tensors, row reductions, the fused GEMM
epilogue, attention tuning, fail-closed benchmark evidence tooling, and
the CI gates the workflow now runs.

Verified on this tree: cargo test --release with and without quant-prep
(392 and 402 tests), clippy -D warnings for both feature sets, rustfmt,
cargo doc, the 2500-case gemm_fuzz_deep soak, gemma-metal (196 tests) and
metal-native (51 tests). Note: both dependents build only with their own
pending, uncommitted adaptations (gemma-metal build.rs and src, metal-native
build.rs and src), which are not part of this commit.

+++ Rust_MLKit/crates/tessl/kernels/gemm_q4_mlx.metal
@@ -73,20 +73,20 @@ kernel void gemm_q4_mlx_simd(
     float acc[SIMD_ROWS][GEMM_MAX_M];
     for (uint r = 0u; r < SIMD_ROWS; ++r)
         for (uint m = 0u; m < GEMM_MAX_M; ++m) acc[r][m] = 0.0f;
-    for (uint k0 = 0u; k0 < cols; k0 += SIMD_BLOCK) {
-        const uint col = k0 + lane_col0;
-        if (col + SIMD_VPT <= cols) {
+    for (ulong k0 = 0ul; k0 < (ulong)cols; k0 += SIMD_BLOCK) {
+        const ulong col = k0 + lane_col0;
+        if (col + SIMD_VPT <= (ulong)cols) {
             float xt[GEMM_MAX_M][16];
             float xsum[GEMM_MAX_M];
             for (uint m = 0u; m < m_cap; ++m)
-                xsum[m] = load_x16_qdot(x + m * cols + col, xt[m]);
-            const uint g = col / group_size;
-            const uint byte_off = col >> 1;
+                xsum[m] = load_x16_qdot(x + (ulong)m * cols + col, xt[m]);
+            const ulong g = col / group_size;
+            const ulong byte_off = col >> 1;
             for (uint r = 0u; r < SIMD_ROWS; ++r) {
                 const uint row = row0 + r;
                 if (row >= rows) break;
-                const bfloat2 sbv = sb[row * gpr + g];
-                device const uchar *wp = packed + row * row_bytes + byte_off;
+                const bfloat2 sbv = sb[(ulong)row * gpr + g];
+                device const uchar *wp = packed + (ulong)row * row_bytes + byte_off;
                 for (uint m = 0u; m < m_cap; ++m)
                     acc[r][m] += qdot16(wp, xt[m], float(sbv.x), float(sbv.y), xsum[m]);
             }
@@ -96,7 +96,7 @@ kernel void gemm_q4_mlx_simd(
         const uint row = row0 + r;
         for (uint m = 0u; m < m_cap; ++m) {
             const float sum = simd_sum(acc[r][m]);
-            if (lane == 0u && row < rows) y[m * rows + row] = sum;
+            if (lane == 0u && row < rows) y[(ulong)m * rows + row] = sum;
         }
     }
 }
@@ -127,20 +127,20 @@ kernel void gemm_q4_mlx_simd_add(
     float acc[SIMD_ROWS][GEMM_MAX_M];
     for (uint r = 0u; r < SIMD_ROWS; ++r)
         for (uint m = 0u; m < GEMM_MAX_M; ++m) acc[r][m] = 0.0f;
-    for (uint k0 = 0u; k0 < cols; k0 += SIMD_BLOCK) {
-        const uint col = k0 + lane_col0;
-        if (col + SIMD_VPT <= cols) {
+    for (ulong k0 = 0ul; k0 < (ulong)cols; k0 += SIMD_BLOCK) {
+        const ulong col = k0 + lane_col0;
+        if (col + SIMD_VPT <= (ulong)cols) {
             float xt[GEMM_MAX_M][16];
             float xsum[GEMM_MAX_M];
             for (uint m = 0u; m < m_cap; ++m)
-                xsum[m] = load_x16_qdot(x + m * cols + col, xt[m]);
-            const uint g = col / group_size;
-            const uint byte_off = col >> 1;
+                xsum[m] = load_x16_qdot(x + (ulong)m * cols + col, xt[m]);
+            const ulong g = col / group_size;
+            const ulong byte_off = col >> 1;
             for (uint r = 0u; r < SIMD_ROWS; ++r) {
                 const uint row = row0 + r;
                 if (row >= rows) break;
-                const bfloat2 sbv = sb[row * gpr + g];
-                device const uchar *wp = packed + row * row_bytes + byte_off;
+                const bfloat2 sbv = sb[(ulong)row * gpr + g];
+                device const uchar *wp = packed + (ulong)row * row_bytes + byte_off;
                 for (uint m = 0u; m < m_cap; ++m)
                     acc[r][m] += qdot16(wp, xt[m], float(sbv.x), float(sbv.y), xsum[m]);
             }
@@ -151,7 +151,7 @@ kernel void gemm_q4_mlx_simd_add(
         for (uint m = 0u; m < m_cap; ++m) {
             const float sum = simd_sum(acc[r][m]);
             if (lane == 0u && row < rows)
-                y[m * rows + row] = sum + resid[m * rows + row];
+                y[(ulong)m * rows + row] = sum + resid[(ulong)m * rows + row];
         }
     }
 }
@@ -182,17 +182,18 @@ kernel void gemm_q4_mlx_simd_i4(
     float acc[SIMD_ROWS][GEMM_MAX_M];
     for (uint r = 0u; r < SIMD_ROWS; ++r)
         for (uint m = 0u; m < GEMM_MAX_M; ++m) acc[r][m] = 0.0f;
-    for (uint k0 = 0u; k0 < cols; k0 += SIMD_BLOCK) {
-        const uint col = k0 + lane_col0;
-        if (col + SIMD_VPT <= cols) {
+    for (ulong k0 = 0ul; k0 < (ulong)cols; k0 += SIMD_BLOCK) {
+        const ulong col = k0 + lane_col0;
+        if (col + SIMD_VPT <= (ulong)cols) {
             float xt[GEMM_MAX_M][16];
             float xsum[GEMM_MAX_M];
             for (uint m = 0u; m < m_cap; ++m)
-                xsum[m] = load_x16_qdot(x + m * cols + col, xt[m]);
-            const uint g = col / group_size;
-            const uint pack2 = col >> 4;
-            device const uchar *wp0 = packed + ((tile * packs_u2 + pack2) * SIMD_ROWS) * 8u;
-            const uint sb0 = (tile * gpr + g) * SIMD_ROWS;
+                xsum[m] = load_x16_qdot(x + (ulong)m * cols + col, xt[m]);
+            const ulong g = col / group_size;
+            const ulong pack2 = col >> 4;
+            device const uchar *wp0 = packed
+                + (((ulong)tile * packs_u2 + pack2) * SIMD_ROWS) * 8ul;
+            const ulong sb0 = ((ulong)tile * gpr + g) * SIMD_ROWS;
             for (uint r = 0u; r < SIMD_ROWS; ++r) {
                 if (row0 + r >= rows) break;
                 const bfloat2 sbv = sb[sb0 + r];
@@ -205,7 +206,7 @@ kernel void gemm_q4_mlx_simd_i4(
         const uint row = row0 + r;
         for (uint m = 0u; m < m_cap; ++m) {
             const float sum = simd_sum(acc[r][m]);
-            if (lane == 0u && row < rows) y[m * rows + row] = sum;
+            if (lane == 0u && row < rows) y[(ulong)m * rows + row] = sum;
         }
     }
 }
@@ -237,17 +238,18 @@ kernel void gemm_q4_mlx_simd_add_i4(
     float acc[SIMD_ROWS][GEMM_MAX_M];
     for (uint r = 0u; r < SIMD_ROWS; ++r)
         for (uint m = 0u; m < GEMM_MAX_M; ++m) acc[r][m] = 0.0f;
-    for (uint k0 = 0u; k0 < cols; k0 += SIMD_BLOCK) {
-        const uint col = k0 + lane_col0;
-        if (col + SIMD_VPT <= cols) {
+    for (ulong k0 = 0ul; k0 < (ulong)cols; k0 += SIMD_BLOCK) {
+        const ulong col = k0 + lane_col0;
+        if (col + SIMD_VPT <= (ulong)cols) {
             float xt[GEMM_MAX_M][16];
             float xsum[GEMM_MAX_M];
             for (uint m = 0u; m < m_cap; ++m)
-                xsum[m] = load_x16_qdot(x + m * cols + col, xt[m]);
-            const uint g = col / group_size;
-            const uint pack2 = col >> 4;
-            device const uchar *wp0 = packed + ((tile * packs_u2 + pack2) * SIMD_ROWS) * 8u;
-            const uint sb0 = (tile * gpr + g) * SIMD_ROWS;
+                xsum[m] = load_x16_qdot(x + (ulong)m * cols + col, xt[m]);
+            const ulong g = col / group_size;
+            const ulong pack2 = col >> 4;
+            device const uchar *wp0 = packed
+                + (((ulong)tile * packs_u2 + pack2) * SIMD_ROWS) * 8ul;
+            const ulong sb0 = ((ulong)tile * gpr + g) * SIMD_ROWS;
             for (uint r = 0u; r < SIMD_ROWS; ++r) {
                 if (row0 + r >= rows) break;
                 const bfloat2 sbv = sb[sb0 + r];
@@ -261,7 +263,7 @@ kernel void gemm_q4_mlx_simd_add_i4(
         for (uint m = 0u; m < m_cap; ++m) {
             const float sum = simd_sum(acc[r][m]);
             if (lane == 0u && row < rows)
-                y[m * rows + row] = sum + resid[m * rows + row];
+                y[(ulong)m * rows + row] = sum + resid[(ulong)m * rows + row];
         }
     }
 }
